+++ source/maxAgreement.c
@@ -4,7 +4,7 @@
 
 BranchArray* treeToBranchModified(Tree* tree, int* permutation)
 {
-    INT p = 1;
+    uint64_t p = 1;
     int i = 0;
     int j = 0;
     unsigned branchNum = tree->nodesNum;
